+++ generator/src/jvmMain/kotlin/tech/skot/tools/generation/ComponentDef.kt
@@ -219,18 +219,18 @@ fun KClass<out SKComponentVC>.def(): ComponentDef {
                 }
                 PropertyDef(
                     it.name,
-                    it.returnType.asTypeName().stripJsAnnotations(),
+                    it.returnType.asCleanTypeName(),
                     meOrSubComponentHasState = (it.returnType.classifier as KClass<out SKComponentVC>).meOrSubComponentHasState(),
                     passToParentView = it.hasAnnotation<SKPassToParentView>(),
                 )
             },
         fixProperties =
             stateProperties.filter { (it !is KMutableProperty) }.map {
-                PropertyDef(name = it.name, type = it.returnType.asTypeName().stripJsAnnotations())
+                PropertyDef(name = it.name, type = it.returnType.asCleanTypeName())
             },
         mutableProperties =
             stateProperties.filterIsInstance<KMutableProperty<*>>().map {
-                PropertyDef(name = it.name, type = it.returnType.asTypeName().stripJsAnnotations())
+                PropertyDef(name = it.name, type = it.returnType.asCleanTypeName())
             },
         state = nestedClasses.find { it.hasAnnotation<SKUIState>() }?.asClassName(),
         ownFunctions = ownFunctions(),