+++ plugin/src/main/kotlin/tech/skot/Versions.kt
@@ -2,11 +2,14 @@
 
 package tech.skot
 
+import kotlin.Int
+import kotlin.String
+
 /**
  * This code is generated from build.gradle.kts
  */
 public object Versions {
-  public const val skot: String = "1.5.3-ua"
+  public const val skot: String = "1.5.4-ua"
 
   public const val group: String = "com.github.useradgents.skot"